+++ typescript/src/tests/day-1.test.ts
@@ -1,6 +1,6 @@
 import path from "path";
-import { default as day1 } from "../solutions/day-01-calorie-counting-1";
-import { default as day2 } from "../solutions/day-01-calorie-counting-2";
+import { default as part1 } from "../solutions/day-01-calorie-counting-1";
+import { default as part2 } from "../solutions/day-01-calorie-counting-2";
 import { readFileSync } from "fs";
 
 describe("day 1", () => {
@@ -10,11 +10,11 @@ describe("day 1", () => {
     input = readFileSync(input_file_name, "utf-8");
   });
   test("part 1", () => {
-    let output = new day1().solve(input);
+    let output = new part1().solve(input);
     expect(output).toBe(24000);
   });
   test("part 2", () => {
-    let output = new day2().solve(input);
+    let output = new part2().solve(input);
     expect(output).toBe(45000);
   });
 });
